% Generated by roxygen2 (4.0.1): do not edit by hand
\name{render_mustache}
\alias{render_mustache}
\title{Render mustache template with data}
\usage{
render_mustache(template, data)
}
\arguments{
\item{template}{character, containing the template}

\item{data}{named list, containing the data}
}
\description{
Render mustache template with data
}

